Add files via upload
The attached bank transactional data can be used to try the code, please note , it has data for only 4 months. 

Results: The results are partial , in the sence that if you do not consider dates ( time). you want to create one set of variables.
However, if you want to run the code for 1,3,6,12 months then you have to give the perameters. to get the respective variables.
Ex: Results are shown to create 5 variables, but if you give months peraments as mentioned above , you will get 5X4 = 20 variables.